Registered Company Name saved to the transient registration (#1140)
https://eaflood.atlassian.net/browse/RUBY-1745

Here we have the work for confirming the registered company name. Once the user has confirmed the name and address that is pulled from the company house API (using their inputted company number) we are then saving their registered company name into the transient registration. This populates in the database as registeredCompanyName. This leaves the field companyName in the database which is optional for the user to enter a trading/business name